+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2018-09-26
+
+### Added
+* upon successful server startup, log:
+  * list of enabled NFS versions
+  * list of exports
+  * list of ports that should be exposed
+* improved error detection and logging
+
 ## [1.1.1] - 2018-08-21
 
 ### Fixed